Fix sporadic failing of get_lca test
Two commits made with the exact same changes and
same details on an empty repository within a
time-tick, have the chance to generate the same
hash (as no parent hash is involved which would
salt the hash with the repositories history).
This is not a bug but a design aspect, and thus
has to be respected in the tests.